refactor: 同意と復元の共有処理をdesktop-runtimeへ抽出
